+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/beans/BeanDefinition.java
@@ -1,4 +1,4 @@
-package com.yrb.tinyioc;
+package com.yrb.tinyioc.beans;
 
 /**
  * @author bjyangrubing