+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/context/ApplicationContext.java
@@ -1,0 +1,13 @@
+package com.yrb.tinyioc.context;
+
+import com.yrb.tinyioc.beans.factory.BeanFactory;
+
+/**
+ * @author bjyangrubing
+ * @createTime 2017/8/7 19:41
+ * Description: 
+ */
+public interface ApplicationContext extends BeanFactory
+{
+
+}